backend/requirements.txt
mistralai
fastapi
uvicorn
python-dotenv
pydantic